requirements.txt
discord.py
python-dotenv
rcon
requests
sqlalchemy